refactor(desktop): extract tray image construction
